requirements.txt
keras==2.14.0
matplotlib==3.8.0
numpy==1.26.1
pandas==2.1.1
Pillow>=9.5.0,<=10.1.0

streamlit==1.27.0
tensorflow==2.14.0
